Added characteristics update logic

+++ sephora-backend/Application/Services/Implementations/ProductService.cs
@@ -2,6 +2,7 @@
 
 public class ProductService(
     IRepository<ProductEntity> productRepository,
+    IRepository<Characteristic> characteristicRepository,
     IPieceService pieceService,
     IFavoritesService favService,
     IMapper mapper
@@ -9,30 +10,64 @@ public class ProductService(
 {
     public async Task Create(CreateProductDto createProductDto)
     {
-        await productRepository.Insert(
-            mapper.Map<ProductEntity>(createProductDto)
-        );
+        var entity = mapper.Map<ProductEntity>(createProductDto);
+        await productRepository.Insert(entity);
         await productRepository.Save();
+
+        foreach (var charDto in createProductDto.Characteristics)
+        {
+            var charEntity = mapper.Map<Characteristic>(charDto);
+            charEntity.ProductId = entity.Id;
+            await characteristicRepository.Insert(charEntity);
+        }
+
+        await characteristicRepository.Save();
     }
 
     public async Task Delete(long id)
     {
+        var product = await productRepository.GetItemBySpec(
+            new Products.GetById(id)
+        );
+        if (product is null)
+            throw new ArgumentException(
+                $"Product with the id={{{id}}} is not found"
+            );
+
         // delete pieces directly so the files are also cleaned up
-        var pieces = (await productRepository.GetItemBySpec(
-                new Products.GetById(id))
-            )?.ProductPieces;
-        foreach (var piece in pieces ?? [])
+        var pieces = product.ProductPieces;
+        foreach (var piece in pieces)
             await pieceService.Delete(mapper.Map<ProductPiece>(piece).Id);
 
-        await productRepository.Delete(id);
+        foreach (var c in product.Characteristics)
+            await characteristicRepository.Delete(c);
+
+        await productRepository.Delete(product);
         await productRepository.Save();
     }
 
     public async Task Edit(EditProductDto editProductDto)
     {
-        await productRepository.Update(
-            mapper.Map<ProductEntity>(editProductDto)
-        );
+        var entity = mapper.Map<ProductEntity>(editProductDto);
+        await productRepository.Update(entity);
+
+        foreach (var c in entity.Characteristics)
+        {
+            c.ProductId = editProductDto.Id;
+            switch (entity.Id)
+            {
+                case 0:
+                    await characteristicRepository.Insert(c);
+                    break;
+                case -1:
+                    await characteristicRepository.Delete(c);
+                    break;
+                default:
+                    await characteristicRepository.Update(c);
+                    break;
+            }
+        }
+
         await productRepository.Save();
     }
